pyproject.toml
[project]
name = "open-range"
version = "0.1.0"
description = "Multi-agent cybersecurity gymnasium built on OpenEnv"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "fastapi>=0.110",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "docker>=7.0",
    "jinja2>=3.1",
    "litellm>=1.30",
    "uvicorn>=0.27",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27"]
training = ["trl>=0.8", "unsloth"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/open_range"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
